drivers: nfc: pn547: Use a unsigned int
* Fixes NFC for bardockpro 
* Need I say more :)

https://github.com/LineageOS/android_system_nfc/blob/lineage-15.1/halimpl/pn54x/tml/phTmlNfc_i2c.c#L287

+++ drivers/nfc/pn547.h
@@ -25,7 +25,7 @@
  */
  
 #define PN547_MAGIC	0xE9
-#define PN547_SET_PWR	_IOW(PN547_MAGIC, 0x01, long)
+#define PN547_SET_PWR	_IOW(PN547_MAGIC, 0x01, unsigned int)
 
 
 #define PN547_NAME "pn544"
